Initial draft of event handler model allowing registration of handlers against the map component.
Changes made to expose zoom as a bidirectional property, and center as a read-only property. Additions to the sample, and an extra test to test addition to the JavascriptObject to allow wrapping of JSObjects coming back from Javascript.

Alterations to send back the MouseEvent object to the UIEventHandler, and implemented tracking UI events on other components, such as the Polyline, which is also implemented.